requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiokafka==0.13.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
black==26.3.1
boto3==1.40.61
botocore==1.40.61
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
coverage==7.13.4
cryptography==46.0.5
deepgram-sdk==6.0.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
librt==0.8.1
limits==5.8.0
mailgun2==2.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
openai==2.26.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
psycopg2-binary==2.9.11
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
resend==2.23.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
ruff==0.15.6
s3transfer==0.14.0
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
yarl==1.23.0
